+++ CHANGES.md
@@ -19,3 +19,4 @@ Changelog
 * Improved logging.
 * Improved documentation.
 * Easier unregistration/shutdown using either `java.lang.Runtime` shutdown hooks or customized shutdown systems (Issue #20).
+* Utilities for shutdown in a servlet environment.